Add CampusModel. Update parsing swipe data mechanism
