+++ app/src/main/java/com/cornellappdev/android/eatery/model/DiningHallModel.java
@@ -23,11 +23,10 @@ import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 
-public class DiningHallModel extends EateryBaseModel implements Serializable {
+public class DiningHallModel extends CampusModel implements Serializable {
 
     private Map<LocalDate, DailyMenuModel> mWeeklyMenu;
     private List<LocalDate> mSortedDates;
-    private List<Swipe> mSwipeDataList;
 
     public static DiningHallModel fromJSON(Context context, boolean hardcoded, JSONObject eatery)
             throws JSONException {
@@ -49,10 +48,6 @@ public class DiningHallModel extends EateryBaseModel implements Serializable {
             today = today.minusDays(1);
         }
         return mWeeklyMenu.get(today);
-    }
-
-    public List<Swipe> getSwipeData() {
-        return mSwipeDataList;
     }
 
     private MealModel getCurrentMeal() {
@@ -157,33 +152,12 @@ public class DiningHallModel extends EateryBaseModel implements Serializable {
         super.parseEatery(context, hardcoded, eatery);
         mWeeklyMenu = new HashMap<>();
         mSortedDates = new ArrayList<>();
-        mSwipeDataList = new ArrayList<>();
         DateTimeFormatter timeFormatter = new DateTimeFormatterBuilder()
                 .parseCaseInsensitive()
                 .appendPattern("h:mma")
                 .toFormatter(Locale.US);
-        DateTimeFormatter dateFormatter = new DateTimeFormatterBuilder()
-                .parseCaseInsensitive()
-                .appendPattern("yyyy-MM-dd")
-                .toFormatter(Locale.US);
-
         mId = eatery.id();
-
-        // TODO: Fix parsing of swipe data times
-        /*for (AllEateriesQuery.SwipeDatum swipeDatum : eatery.swipeData()) {
-            LocalTime start = null, end = null;
-            start = LocalTime.parse(swipeDatum.startTime().toUpperCase().replaceAll("\\s+", ""),
-                    timeFormatter);
-            end = LocalTime.parse(swipeDatum.endTime().toUpperCase().replaceAll("\\s+", ""),
-                    timeFormatter);
-            mSwipeDataList.add(
-                    new Swipe(start, end, swipeDatum.swipeDensity(), swipeDatum.waitTimeLow(),
-                            swipeDatum.waitTimeHigh()));
-
-        }
-        Collections.sort(mSwipeDataList);
-        */
-
+        parseSwipeData(eatery.swipeData());
         // Each Operating Hour is a single day for dining halls
         for (AllEateriesQuery.OperatingHour operatingHour : eatery.operatingHours()) {
             DailyMenuModel dailyMenuModel = new DailyMenuModel();
